toolchain/linux-musl: Remove GITREV variable
This is the only recipe of all four `litex-conda-*` repositories that
uses `GITREV`. In addition, it is used to set `version` that will be
overwritten by `conda-build-prepare` since the repository has a git
repository as a source.

Because of that the version was set to be `0.0` and the comment was
added about how the version will be established.